vlm_mock/requirements.txt
flask==3.0.3
pytesseract==0.3.13
Pillow==10.4.0
